refactor(workstation): remove vestigial dockFilter machinery
The unified WorkStation tab pool makes the visible content host follow the
active tab via `activeHostAtom`; `dockFilter` no longer drives host
selection, so remove the whole concept without changing behavior.

- Relocate `activeHostAtom` to `store/workstation/tabHost.ts`; re-export it
  from the store barrel so imports keep working.
- Delete the `store/workstation/dockFilter/` module (dockFilterAtom,
  DockFilter type, DEFAULT_DOCK_FILTER, path/query helpers).
- Delete `useDockFilterUrlSync` (the `?filter=` URL sync) and the now-inert
  `useActiveTabHostReconciliation` (activeHost === effectiveHost always).
- Rename `useAppShellDockFilterSync` -> `useAppShellRouteSync`, keeping only
  the base-path side effects (My Station, chat visibility, chat width).
- `useAppShellDock` now always seeds ["code", "browser"] + the active tab's
  host so "New Browser" works and the host still mounts; drop the dead
  handleDockClick / activeDockId.
- Sweep remaining reads in tab list, focus, launch actions, the focused-chat
  rail, tutorials, view service, open-in-workstation helpers, e2e helpers,
  and remove the unused i18n `workstation.dockFilter` keys across locales.

Pre-commit hook ran. Total eslint: 0, total circular: 0
