+++ src/i18n/locales/zh-Hant/navigation.json
@@ -428,13 +428,6 @@
     "title": "探索"
   },
   "workstation": {
-    "dockFilter": {
-      "all": "所有分頁",
-      "code": "程式碼",
-      "browser": "瀏覽器",
-      "data": "資料庫",
-      "project": "專案"
-    },
     "agentComputer": {
       "label": "{{agent}} 的工作站",
       "agents": {
