+++ src/modules/WorkStation/AppShell/hooks/index.ts
@@ -1,8 +1,8 @@
 export * from "./useAppShellActions";
 export * from "./useAppShellDerivedState";
 export * from "./useAppShellDock";
-export * from "./useAppShellDockFilterSync";
 export * from "./useAppShellRepo";
+export * from "./useAppShellRouteSync";
 export * from "./useAppShellSimulatorPanelSync";
 export * from "./useAppShellStationMode";
 export * from "./useAppShellStatusBar";
