+++ src/modules/index.tsx
@@ -66,7 +66,6 @@ import {
   sessionChatPositionAtom,
   workStationChatPositionAtom,
 } from "@src/store/ui/workStationAtom";
-import { dockFilterAtom } from "@src/store/workstation";
 import { prewarmColor } from "@src/util/ui/theme/glassMaterial";
 
 import { BackgroundLayer } from "./shared/components";
@@ -253,7 +252,6 @@ const AppShell = () => {
   const setStationMode = useSetAtom(stationModeAtom);
   const setSidebarCollapsed = useSetAtom(sidebarCollapsedAtom);
   const setStationChatVisibility = useSetAtom(stationChatVisibilityAtom);
-  const setDockFilter = useSetAtom(dockFilterAtom);
   const [tutorialsModalOpen, setTutorialsModalOpen] = useState(false);
   const [generalLayoutTourOpen, setGeneralLayoutTourOpen] = useState(false);
   const [generalLayoutTourRunId, setGeneralLayoutTourRunId] = useState(0);
@@ -287,7 +285,6 @@ const AppShell = () => {
       "my-station": true,
     }));
     restoreChatWidth();
-    setDockFilter("all");
     setGeneralLayoutTourRunId((value) => value + 1);
     window.setTimeout(() => setGeneralLayoutTourOpen(true), 220);
   }, [
@@ -295,7 +292,6 @@ const AppShell = () => {
     navigate,
     restoreChatWidth,
     setChatPanelMaximized,
-    setDockFilter,
     setSidebarCollapsed,
     setStationChatVisibility,
     setStationMode,
@@ -314,7 +310,6 @@ const AppShell = () => {
       "my-station": true,
     }));
     restoreChatWidth();
-    setDockFilter("code");
     setCodeEditorTourRunId((value) => value + 1);
     window.setTimeout(() => setCodeEditorTourOpen(true), 240);
   }, [
@@ -322,7 +317,6 @@ const AppShell = () => {
     navigate,
     restoreChatWidth,
     setChatPanelMaximized,
-    setDockFilter,
     setSidebarCollapsed,
     setStationChatVisibility,
     setStationMode,
