--- src/store/workstation/dockFilter/index.ts
@@ -1,11 +1,0 @@
-export {
-  type DockFilter,
-  DEFAULT_DOCK_FILTER,
-  activeHostAtom,
-  dockFilterAtom,
-} from "./atoms";
-export {
-  DOCK_FILTER_QUERY_KEY,
-  buildDockFilterPath,
-  composeDockFilterSearch,
-} from "./navigation";
